# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/orders": {
      "post": {
        "tags": [
          "api"
        ],
        "summary": "Create Order",
        "description": "Create a new order from website.\n\nAccepts customer information and cart items. Creates or updates customer record,\ncreates order with NEW_LEAD status and lead_source=SITE.\n\nReturns created order details.",
        "operationId": "create_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/manager/orders": {
      "get": {
        "tags": [
          "manager-orders"
        ],
        "summary": "Get Manager Orders",
        "operationId": "get_manager_orders",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "segment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(b2c|b2b)$",
              "default": "b2c",
              "title": "Segment"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "overdue_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue Only"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at_desc",
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerOrderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/manager/orders/{order_id}": {
      "get": {
        "tags": [
          "manager-orders"
        ],
        "summary": "Get Manager Order Detail",
        "operationId": "get_manager_order_detail",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerOrderDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "manager-orders"
        ],
        "summary": "Patch Manager Order",
        "operationId": "patch_manager_order",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerOrderUpdatePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerOrderDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/manager/orders/{order_id}/documents/{doc_type}": {
      "post": {
        "tags": [
          "manager-orders"
        ],
        "summary": "Generate Manager Order Document",
        "operationId": "generate_manager_order_document",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          },
          {
            "name": "doc_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagerOrderDocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/manager/{full_path}": {
      "get": {
        "summary": "Serve Manager App",
        "operationId": "serve_manager_app_manager__full_path__get",
        "parameters": [
          {
            "name": "full_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Full Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CartItemPayload": {
        "properties": {
          "product_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity",
            "default": 1
          },
          "with_installation": {
            "type": "boolean",
            "title": "With Installation",
            "default": false
          },
          "installation_price": {
            "type": "number",
            "title": "Installation Price",
            "default": 0.0
          },
          "installation_meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Meta"
          },
          "installation_options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Options",
            "default": []
          }
        },
        "type": "object",
        "title": "CartItemPayload"
      },
...
      "CustomerPayload": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": "individual"
          },
          "full_legal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Legal Name"
          },
          "inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inn"
          },
          "legal_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legal Address"
          },
          "iban": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iban"
          },
          "bic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bic"
          },
          "bank_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank Name"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone"
        ],
        "title": "CustomerPayload"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ManagerOrderDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "next_followup_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Followup Date"
          },
          "assessment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Date"
          },
          "installation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Date"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "margin": {
            "type": "number",
            "title": "Margin"
          },
          "is_paid": {
            "type": "boolean",
            "title": "Is Paid"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "delivery_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Address"
          },
          "customer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderCustomerBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "product_lines": {
            "items": {
              "$ref": "#/components/schemas/OrderProductLineResponse"
            },
            "type": "array",
            "title": "Product Lines",
            "default": []
          },
          "service_lines": {
            "items": {
              "$ref": "#/components/schemas/OrderServiceLineResponse"
            },
            "type": "array",
            "title": "Service Lines",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "total_amount",
          "total_cost",
          "margin",
          "is_paid"
        ],
        "title": "ManagerOrderDetailResponse"
      },
      "ManagerOrderDocumentResponse": {
        "properties": {
          "doc_id": {
            "type": "integer",
            "title": "Doc Id"
          },
          "doc_type": {
            "type": "string",
            "title": "Doc Type"
          },
          "edit_url": {
            "type": "string",
            "title": "Edit Url"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "doc_type",
          "edit_url"
        ],
        "title": "ManagerOrderDocumentResponse"
      },
      "ManagerOrderListItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "next_followup_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Followup Date"
          },
          "assessment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Date"
          },
          "installation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Date"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "total_cost": {
            "type": "number",
            "title": "Total Cost"
          },
          "margin": {
            "type": "number",
            "title": "Margin"
          },
          "is_paid": {
            "type": "boolean",
            "title": "Is Paid"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "delivery_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Address"
          },
          "customer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderCustomerBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "created_at",
          "total_amount",
          "total_cost",
          "margin",
          "is_paid"
        ],
        "title": "ManagerOrderListItemResponse"
      },
      "ManagerOrderListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ManagerOrderListItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "ManagerOrderListResponse"
      },
...
      "ManagerOrderUpdatePayload": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "next_followup_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Followup Date"
          },
          "assessment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Date"
          },
          "installation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installation Date"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "is_paid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Paid"
          },
          "products": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ManagerOrderProductLinePayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Products"
          },
          "services": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ManagerOrderServiceLinePayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Services"
          }
        },
        "type": "object",
        "title": "ManagerOrderUpdatePayload"
      },
      "Meta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "Meta"
      },
      "OrderCustomerBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "full_legal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Legal Name"
          },
          "inn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inn"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "phone"
        ],
        "title": "OrderCustomerBrief"
      },
      "OrderPayload": {
        "properties": {
          "customer": {
            "$ref": "#/components/schemas/CustomerPayload"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CartItemPayload"
            },
            "type": "array",
            "title": "Items",
            "default": []
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "customer"
        ],
        "title": "OrderPayload"
      },
...
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "total_amount",
          "created_at"
        ],
        "title": "OrderResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/login/access-token"
          }
        }
      }
    }
  }
}